+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support pull requests - by Awiteb
 - Add `--personal-fork` flag to `repo announce` command - by Awiteb
 - Log to stderr and a file - by Awiteb
+- Accept patches from stdin in `patch send` command - by Awiteb
 
 ### Breaking Change